requirements.txt
fpdf==1.7.2
Pillow==10.1.0
pynput==1.7.6
